Add value to hidden_field
I added a value (:book) to the hidden_field method because the field stays blank as shown by the validation line under. But when adding :book as the value, the book name stays as "book" as opposed to the actual book title that is passed in from the <a> tag.